Convert Images and Illustrations in epub
Should now be able to also download images in chapters (frequent in Japanese LNs) and put them in the epub.
Can't catch the cover image of the epub yet though, will probably be easy if sites have set html pages like wuxiaworld or royalroad, but for the variety of all wordpress and blogspot blog like sites, it's a little harder. As a matter of fact, metadata for these two sites are pretty much nonexistant, as they are different for each site (html wise).